# api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Boost rest APi",
    "version": "2.0",
    "title": "boost rest api",
    "contact": {
      "name": "Ats dev team"
    },
    "license": {}
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/apiAfin/students": {
      "get": {
        "tags": [
          "student-resource"
        ],
        "summary": "find all students by page",
        "operationId": "getStudentPaginateUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "size",
            "description": "size",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "in": "body",
            "name": "page",
            "description": "page",
            "required": false,
            "schema": {
              "type": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«Student»"
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/apiAfin/students/": {
      "post": {
        "tags": [
          "student-resource"
        ],
        "summary": "create new Student",
        "operationId": "createStudentUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "student",
            "description": "student",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "student-resource"
        ],
        "summary": "update students",
        "operationId": "updateStudentUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "student",
            "description": "student",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/apiAfin/students/addcourse": {
      "put": {
        "tags": [
          "student-resource"
        ],
        "summary": "find add new course to student courses list",
        "operationId": "addStudentCoursesUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "s_id",
            "description": "s_id",
            "required": false,
            "schema": {
              "type": "long"
            }
          },
          {
            "in": "body",
            "name": "c_id",
            "description": "c_id",
            "required": false,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/apiAfin/students/all": {
      "get": {
        "tags": [
          "student-resource"
        ],
        "summary": "find all students",
        "operationId": "getAllStudentUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Student"
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/apiAfin/students/{id}": {
      "get": {
        "tags": [
          "student-resource"
        ],
        "summary": "find student by id",
        "operationId": "getStudentUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/apiAfin/teachers/{id}/course": {
      "get": {
        "tags": [
          "teacher-resource"
        ],
        "summary": "find teachers courses id param is for teacher",
        "operationId": "getCoursesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Course"
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "LocalDateTime": {
      "properties": {
        "chronology": {
          "$ref": "#/definitions/Chronology"
        },
        "dayOfMonth": {
          "type": "integer",
          "format": "int32"
        },
        "dayOfWeek": {
          "type": "string",
          "enum": [
            "MONDAY",
            "TUESDAY",
            "WEDNESDAY",
            "THURSDAY",
            "FRIDAY",
            "SATURDAY",
            "SUNDAY"
          ]
        },
        "dayOfYear": {
          "type": "integer",
          "format": "int32"
        },
        "hour": {
          "type": "integer",
          "format": "int32"
        },
        "minute": {
          "type": "integer",
          "format": "int32"
        },
        "month": {
          "type": "string",
          "enum": [
            "JANUARY",
            "FEBRUARY",
            "MARCH",
            "APRIL",
            "MAY",
            "JUNE",
            "JULY",
            "AUGUST",
            "SEPTEMBER",
            "OCTOBER",
            "NOVEMBER",
            "DECEMBER"
          ]
        },
        "monthValue": {
          "type": "integer",
          "format": "int32"
        },
        "nano": {
          "type": "integer",
          "format": "int32"
        },
        "second": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "ModuleMaterial": {
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Marks": {
      "properties": {
        "hwQuestion": {
          "$ref": "#/definitions/HwQuestion"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "score": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "LocalDate": {
      "properties": {
        "chronology": {
          "$ref": "#/definitions/IsoChronology"
        },
        "dayOfMonth": {
          "type": "integer",
          "format": "int32"
        },
        "dayOfWeek": {
          "type": "string",
          "enum": [
            "MONDAY",
            "TUESDAY",
            "WEDNESDAY",
            "THURSDAY",
            "FRIDAY",
            "SATURDAY",
            "SUNDAY"
          ]
        },
        "dayOfYear": {
          "type": "integer",
          "format": "int32"
        },
        "era": {
          "$ref": "#/definitions/Era"
        },
        "leapYear": {
          "type": "boolean"
        },
        "month": {
          "type": "integer",
          "format": "int32"
        },
        "monthValue": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ZoneOffset": {
      "properties": {
        "id": {
          "type": "string"
        },
        "rules": {
          "$ref": "#/definitions/ZoneRules"
        },
        "totalSeconds": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Solutions": {
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "registering": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "text": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "ZonedDateTime": {
      "properties": {
        "chronology": {
          "$ref": "#/definitions/Chronology"
        },
        "dayOfMonth": {
          "type": "integer",
          "format": "int32"
        },
        "dayOfWeek": {
          "type": "string",
          "enum": [
            "MONDAY",
            "TUESDAY",
            "WEDNESDAY",
            "THURSDAY",
            "FRIDAY",
            "SATURDAY",
            "SUNDAY"
          ]
        },
        "dayOfYear": {
          "type": "integer",
          "format": "int32"
        },
        "hour": {
          "type": "integer",
          "format": "int32"
        },
        "minute": {
          "type": "integer",
          "format": "int32"
        },
        "month": {
          "type": "string",
          "enum": [
            "JANUARY",
            "FEBRUARY",
            "MARCH",
            "APRIL",
            "MAY",
            "JUNE",
            "JULY",
            "AUGUST",
            "SEPTEMBER",
            "OCTOBER",
            "NOVEMBER",
            "DECEMBER"
          ]
        },
        "monthValue": {
          "type": "integer",
          "format": "int32"
        },
        "nano": {
          "type": "integer",
          "format": "int32"
        },
        "offset": {
          "$ref": "#/definitions/ZoneOffset"
        },
        "second": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        },
        "zone": {
          "$ref": "#/definitions/ZoneId"
        }
      }
    },
...
    "CourseModule": {
      "properties": {
        "homework": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Homework"
          }
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "moduleMaterials": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ModuleMaterial"
          }
        },
        "opening": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "sommary": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "Sort": {},
    "Chronology": {
      "properties": {
        "calendarType": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "ZoneOffsetTransition": {
      "properties": {
        "dateTimeAfter": {
          "$ref": "#/definitions/LocalDateTime"
        },
        "dateTimeBefore": {
          "$ref": "#/definitions/LocalDateTime"
        },
        "duration": {
          "$ref": "#/definitions/Duration"
        },
        "gap": {
          "type": "boolean"
        },
        "instant": {
          "$ref": "#/definitions/Instant"
        },
        "overlap": {
          "type": "boolean"
        }
      }
    },
    "Course": {
      "properties": {
        "closing": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "courseModules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CourseModule"
          }
        },
        "discussions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Discussion"
          }
        },
        "domain": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "opening": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "registring": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "teachers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Teacher"
          }
        },
        "title": {
          "type": "string"
        }
      }
    },
    "ZoneOffsetTransitionRule": {
      "properties": {
        "dayOfMonthIndicator": {
          "type": "integer",
          "format": "int32"
        },
        "dayOfWeek": {
          "type": "string",
          "enum": [
            "MONDAY",
            "TUESDAY",
            "WEDNESDAY",
            "THURSDAY",
            "FRIDAY",
            "SATURDAY",
            "SUNDAY"
          ]
        },
        "localTime": {
          "$ref": "#/definitions/LocalTime"
        },
        "midnightEndOfDay": {
          "type": "boolean"
        },
        "month": {
          "type": "string",
          "enum": [
            "JANUARY",
            "FEBRUARY",
            "MARCH",
            "APRIL",
            "MAY",
            "JUNE",
            "JULY",
            "AUGUST",
            "SEPTEMBER",
            "OCTOBER",
            "NOVEMBER",
            "DECEMBER"
          ]
        },
        "timeDefinition": {
          "type": "string",
          "enum": [
            "UTC",
            "WALL",
            "STANDARD"
          ]
        }
      }
    },
...
    "HwQuestion": {
      "properties": {
        "answer": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "proposition1": {
          "type": "string"
        },
        "proposition2": {
          "type": "string"
        },
        "proposition3": {
          "type": "string"
        },
        "proposition4": {
          "type": "string"
        },
        "qmark": {
          "type": "number",
          "format": "double"
        },
        "statement": {
          "type": "string"
        }
      }
    },
    "ZoneId": {
      "properties": {
        "id": {
          "type": "string"
        },
        "rules": {
          "$ref": "#/definitions/ZoneRules"
        }
      }
    },
    "LocalTime": {
      "properties": {
        "hour": {
          "type": "integer",
          "format": "int32"
        },
        "minute": {
          "type": "integer",
          "format": "int32"
        },
        "nano": {
          "type": "integer",
          "format": "int32"
        },
        "second": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "TemporalUnit": {
      "properties": {
        "dateBased": {
          "type": "boolean"
        },
        "durationEstimated": {
          "type": "boolean"
        },
        "timeBased": {
          "type": "boolean"
        }
      }
    },
    "Instant": {
      "properties": {
        "epochSecond": {
          "type": "integer",
          "format": "int64"
        },
        "nano": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Teacher": {
      "properties": {
        "birthday": {
          "$ref": "#/definitions/LocalDate"
        },
        "contact": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "level": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "profession": {
          "type": "string"
        },
        "profile": {
          "type": "string"
        },
        "registring": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "surnames": {
          "type": "string"
        }
      }
    },
    "Duration": {
      "properties": {
        "nano": {
          "type": "integer",
          "format": "int32"
        },
        "negative": {
          "type": "boolean"
        },
        "seconds": {
          "type": "integer",
          "format": "int64"
        },
        "units": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TemporalUnit"
          }
        },
        "zero": {
          "type": "boolean"
        }
      }
    },
    "IsoChronology": {
      "properties": {
        "calendarType": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
...
    "Student": {
      "properties": {
        "birthday": {
          "$ref": "#/definitions/LocalDate"
        },
        "contact": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "courses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Course"
          }
        },
        "email": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "level": {
          "type": "string"
        },
        "marks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Marks"
          }
        },
        "name": {
          "type": "string"
        },
        "profession": {
          "type": "string"
        },
        "profile": {
          "type": "string"
        },
        "registring": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "surnames": {
          "type": "string"
        }
      }
    },
    "Era": {
      "properties": {
        "value": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "ZoneRules": {
      "properties": {
        "fixedOffset": {
          "type": "boolean"
        },
        "transitionRules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ZoneOffsetTransitionRule"
          }
        },
        "transitions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ZoneOffsetTransition"
          }
        }
      }
    },
...
    "Discussion": {
      "properties": {
        "ask": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "registering": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "solutions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Solutions"
          }
        },
        "username": {
          "type": "string"
        }
      }
    },
    "Page«Student»": {
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Student"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Homework": {
      "properties": {
        "hwQuestions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/HwQuestion"
          }
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "score": {
          "type": "number",
          "format": "double"
        }
      }
    }
  }
}